Record snapshots
